+++ Public/Azure/Runs/Receive-AzureThreadRun.ps1
@@ -1,0 +1,57 @@
+function Receive-AzureThreadRun {
+    [CmdletBinding()]
+    [OutputType([pscustomobject])]
+    param (
+        [Parameter(Mandatory = $true, Position = 0, ValueFromPipeline = $true)]
+        [ValidateScript({ ([string]$_.id).StartsWith('run_') -and ([string]$_.thread_id).StartsWith('thread_') })]
+        [Alias('Run')]
+        [Object]$InputObject,
+
+        [Parameter()]
+        [int]$TimeoutSec = 0,
+
+        [Parameter()]
+        [ValidateRange(0, 100)]
+        [int]$MaxRetryCount = 0,
+
+        [Parameter()]
+        [System.Uri]$ApiBase,
+
+        [Parameter()]
+        [string]$ApiVersion,
+
+        [Parameter()]
+        [ValidateSet('azure', 'azure_ad')]
+        [string]$AuthType = 'azure',
+
+        [Parameter()]
+        [securestring][SecureStringTransformation()]$ApiKey,
+
+        [Parameter()]
+        [switch]$Wait,
+
+        [Parameter()]
+        [switch]$AutoRemoveThread
+    )
+
+    begin {
+        # Construct parameters
+        $Parameters = $PSBoundParameters
+        $Parameters.ApiType = [OpenAIApiType]::Azure
+        $Parameters.AuthType = $AuthType
+
+        # Invoke
+        $steppablePipeline = {
+            PSOpenAI\Receive-ThreadRun @Parameters
+        }.GetSteppablePipeline($myInvocation.CommandOrigin)
+        $steppablePipeline.Begin($PSCmdlet)
+    }
+
+    process {
+        $steppablePipeline.Process($PSItem)
+    }
+
+    end {
+        $steppablePipeline.End()
+    }
+}