+++ Public/Azure/Threads/Set-AzureThread.ps1
@@ -1,0 +1,59 @@
+function Set-AzureThread {
+    [CmdletBinding()]
+    [OutputType([pscustomobject])]
+    param (
+        [Parameter(Mandatory = $true, Position = 0, ValueFromPipeline = $true, ValueFromPipelineByPropertyName = $true)]
+        [Alias('thread_id')]
+        [Alias('Thread')]
+        [ValidateScript({
+            ($_ -is [string] -and $_.StartsWith('thread_')) -or `
+                ($_.id -is [string] -and $_.id.StartsWith('thread_')) -or `
+                ($_.thread_id -is [string] -and $_.thread_id.StartsWith('thread_'))
+            })]
+        [Object]$InputObject,
+
+        [Parameter()]
+        [System.Collections.IDictionary]$MetaData,
+
+        [Parameter()]
+        [int]$TimeoutSec = 0,
+
+        [Parameter()]
+        [ValidateRange(0, 100)]
+        [int]$MaxRetryCount = 0,
+
+        [Parameter()]
+        [System.Uri]$ApiBase,
+
+        [Parameter()]
+        [string]$ApiVersion,
+
+        [Parameter()]
+        [ValidateSet('azure', 'azure_ad')]
+        [string]$AuthType = 'azure',
+
+        [Parameter()]
+        [securestring][SecureStringTransformation()]$ApiKey
+    )
+
+    begin {
+        # Construct parameters
+        $Parameters = $PSBoundParameters
+        $Parameters.ApiType = [OpenAIApiType]::Azure
+        $Parameters.AuthType = $AuthType
+
+        # Invoke
+        $steppablePipeline = {
+            PSOpenAI\Set-Thread @Parameters
+        }.GetSteppablePipeline($myInvocation.CommandOrigin)
+        $steppablePipeline.Begin($PSCmdlet)
+    }
+
+    process {
+        $steppablePipeline.Process($PSItem)
+    }
+
+    end {
+        $steppablePipeline.End()
+    }
+}