+++ Public/ConvertTo-Token.ps1
@@ -10,7 +10,7 @@ function ConvertTo-Token {
         [string][LowerCaseTransformation()]$Model,
 
         [Parameter(Mandatory = $false, Position = 1, ParameterSetName = 'encoding')]
-        [ValidateSet('cl100k_base', 'p50k_base', 'p50k_edit', 'r50k_base', 'gpt2')]
+        [ValidateSet('cl100k_base')]
         [string]$Encoding = 'cl100k_base'
     )
 
@@ -30,10 +30,6 @@ function ConvertTo-Token {
 
         $Encoder = switch ($Encoding) {
             'cl100k_base' { [PSOpenAI.TokenizerLib.Cl100kBaseTokenizer]::Encode }
-            'p50k_base' { [PSOpenAI.TokenizerLib.P50kBaseTokenizer]::Encode }
-            'p50k_edit' { [PSOpenAI.TokenizerLib.P50kEditTokenizer]::Encode }
-            'r50k_base' { [PSOpenAI.TokenizerLib.R50kBaseTokenizer]::Encode }
-            'gpt2' { [PSOpenAI.TokenizerLib.Gpt2Tokenizer]::Encode }
         }
     }