+++ Public/Runs/Start-ThreadRun.ps1
@@ -29,10 +29,13 @@ function Start-ThreadRun {
             'gpt-3.5-turbo-0613',
             'gpt-3.5-turbo-16k-0613',
             'gpt-3.5-turbo-1106',
+            'gpt-3.5-turbo-0125',
             'gpt-4-0613',
             'gpt-4-32k',
             'gpt-4-32k-0613',
+            'gpt-4-turbo-preview',
             'gpt-4-1106-preview',
+            'gpt-4-0125-preview',
             'gpt-4-vision-preview'
         )]
         [string][LowerCaseTransformation()]$Model = 'gpt-3.5-turbo',
@@ -97,7 +100,7 @@ function Start-ThreadRun {
 
         # Get API endpoint
         if ($ApiType -eq [OpenAIApiType]::Azure) {
-            $OpenAIParameter = Get-AzureOpenAIAPIEndpoint -EndpointName 'Runs' -Engine $Model -ApiBase $ApiBase -ApiVersion $ApiVersion
+            $OpenAIParameter = Get-AzureOpenAIAPIEndpoint -EndpointName 'Runs' -ApiBase $ApiBase -ApiVersion $ApiVersion
         }
         else {
             $OpenAIParameter = Get-OpenAIAPIEndpoint -EndpointName 'Runs' -ApiBase $ApiBase